将所有与username有关的文件路径命名全部放到BaseSpider中
